gps.c: When there is no fix available, put the receiver in "RCVR_UP" mode
If this is not done, a "Searching for a GPS receiver" notification is
shown to user. However, this is not the case. We are not searching for
the receiver, we are trying to acquire a GPS fix. Changing the
receiver state correctly fixes this problem. The map mark is also kept
moving, in case there is location information available from network
positioning even though we don't have a GPS fix yet.

Signed-off-by: Timo Kokkonen <kaapeli@itanic.dy.fi>