Synchronization of repository and tile source with external XML file.
Implements handler of Sync button in repositories management dialog. It downloads two XML files with
tile sources and repositories settings and performs sync of local settings with that data. The sync
is performed using only some fields, such as URL, name and so on. Settings which are local-specific
are left untouched (for example, refresh interfal of layers, visibility flag, etc).