Optimize drawing of path for small magnifications
Render only those segments which have a visible effect at the current zoom
level.

The map_path_optimize() assigns a zoom value to every point in the path: it
indicates what is the zoom level after which the point segment should be drawn.

The algorythm works like this: for every point P in the path (other than the
first one, which has a zoom value of +127), set n to 1 and

a) calculate the distance D (*) between P and P-n, in map units

b) compute the greatest (**) zoom level Z after which D becomes wider than a certain amount of pixels (3 + _draw_width)

c) search backwards for the previous point P-n having been assigned a zoom value bigger than Z

d) if this point is the one just before P, go to (f)

e) recursively iterate steps from (a) to (e), with the new value of n.

(f) assign Z as zoom value for P

(*) Instead of calculating the precise distance, we simply take the maximum distance in either the X or Y axes -- we don't need any more precision.

(**) "smallest" as numeric value. In maemo-mapper, a small zoom level means bigger magnification; here we use the word "zoom" referring to the numeric value in maemo-mapper code.