New path methods, add indexes for path breaks
The main point of this commit is to add a way to quickly iterate the path lines
(corresponding to GPX trkseg), via the MapLineIter iterators.
From this follows that now there are no NULL points into the paths.

The immediate benefit is that now the track time on the information panel is
correctly calculated, when more than one track is loaded.

Also, several methods and macros have been added to handle the path data,
should we change its internal structure in the future.