map_path_merge(): Don't add breaks to merged paths
This patch removes the first line marker on the appended path, to avoid adding
empty lines.
So the behaviour is:
a) If the first path ends with a break, the two paths will appear as separate
b) If the first path doesn't end with a break, the two paths will appear as one

Also, initialize the path after unsetting it (in the rare case where a NULL
path is substituted).