Deliver map changes via MapEvent
This allows the LayerGroup and the layers to react to map changes in the
same QEventLoop iteration, avoiding therefore visual artifacts due to
different timing.
This is also more efficient (especially when more layers will be
enabled), because signal connections are reduced, and no event is
emitted if nothing has changed (previously, the Layer class was checking
this).

In the same commit, the Makefile rules for generating moc files are also
changed.