Win32: Don't use QGLWidget
On my Lenovo Ideapad s10-3t QGLWidget works fine when no graphics driver
are installed; but with graphics driver, nothing is painted in the
window.
This might be a bug somewhere in mappero; Qt demos and example appear to
work all fine, and QGLWidget works fine in mappero too, if set from the
View::switchFullscreen() method.