Point, Map, PoiView: consistently use QPointF for animations
Add a Point::toPixelF() method to avoid rounding discrepancies, and use
it in Map and PoiView. This helps reducing the jagging of map movements.

Also, when the animatedCenterUnit changes in the Map, treat the final
point (animatedCenterUnit == requestedCenterUnit) as a special case, and
optimize it so that setCenter() is called instead (and only once).