Router: wait for the bottomPanel animation to complete
This is a workaround to the problem that if we emit currentItemChanged
as soon as the route is found, map.lookAt() will use the full-screen
size of the map to compute the zoom level; but at the same time the
bottomPanel animation is running, causing the map to shrink and possibly
cut out parts of the route.
Therefore, we wait until the bottomPanel animation has finished, so that
the map size won't change.