Types: register Utils type as Mappero, Controller as Controller
Switch to use the functions and properties declared in Utils as the new
Mappero object. Add the Controller class as "Controller", and update the
code which was using it.

Update the UIConstants files to use the uiScale property from the Utils
class, and avoid defining the same variables in both of them. Make the
one under src/qt/ include the one provided by libMapperoUi.