Application, PluginManager: improve handling of uninstalled plugins
When building with QBS, the application binary will not be found under
src/qt/, but under bin/ (in Linux, at least); adapt the code to this new
location, and also use QStandardPaths::locate() to achieve more robust
resolution. Let the Application class play with XDG_DATA_DIRS so that
all resources will be found when running uninstalled, too.