Build: change application name and Android package name
Change the application name so that it appears nicer in Android (though
we could also use the `Android:label` attribute in the XML manifest),
and the package name. For some reason, these changes alone trigger a
warning at installation time about the app being developed for an older
version of Android. So let's add version information in the manifest.